Correccion insert.sql y cambios a tiempo de vida

+++ src/main/java/edu/eci/pdsw/entities/EquipoComplejo.java
@@ -22,7 +22,7 @@ public class EquipoComplejo implements Comparable<EquipoComplejo> {
     private String marca;
     private Modelo modelo_Eq;
     private DatosGenerales ordenCompra;
-    private int tiempoRestante;
+    private long tiempoRestante;
     public static String diario = "prestamo diario", p24h = "prestamo 24 horas", mantenimiento = "mantenimiento", almacen = "en almacen", indefinido = "prestamo indefinido", semestre = "prestamo por semestre";
     public static String baja = "dado de baja", reparacion = "en reparacion";
 
@@ -37,7 +37,7 @@ public class EquipoComplejo implements Comparable<EquipoComplejo> {
      * @param tiempoDeUso cuanto tiempo ha sido usado el equipo
      * @throws EquipoException
      */
-    public EquipoComplejo(Modelo mod, String mar, String ser, String plac, DatosGenerales dg, int tiempoDeUso) throws EquipoException {
+    public EquipoComplejo(Modelo mod, String mar, String ser, String plac, DatosGenerales dg, long tiempoDeUso) throws EquipoException {
         if (mod == null) {
             throw new EquipoException(EquipoException.EQUIPOC_SIN_MODELO);
         }
@@ -79,7 +79,7 @@ public class EquipoComplejo implements Comparable<EquipoComplejo> {
      * @param tiempoDeUso e tiempo que ha sido usado el equipo
      * @throws EquipoException
      */
-    public EquipoComplejo(boolean asegurado, boolean disponibilidad, String estado, String serial, String placa, String marca, Modelo modelo_Eq, DatosGenerales dg, int tiempoDeUso) throws EquipoException {
+    public EquipoComplejo(boolean asegurado, boolean disponibilidad, String estado, String serial, String placa, String marca, Modelo modelo_Eq, DatosGenerales dg, long tiempoDeUso) throws EquipoException {
         if (modelo_Eq == null) {
             throw new EquipoException(EquipoException.EQUIPOC_SIN_MODELO);
         }
@@ -121,11 +121,11 @@ public class EquipoComplejo implements Comparable<EquipoComplejo> {
         this.ordenCompra = ordenCompra;
     }
 
-    public int getTiempoRestante() {
+    public long getTiempoRestante() {
         return tiempoRestante;
     }
 
-    public void setTiempoRestante(int tiempoRestante) {
+    public void setTiempoRestante(long tiempoRestante) {
         this.tiempoRestante = tiempoRestante;
     }
